feat (build system): require explicit Kconfig path
Previously the path to .config file was derived from the edt.pickle
path.
In sysbuild we want to be able to access the sysbuild Kconfig namespace.
However, the sysbuild run does not generate the edt.pickle file.
Therefore it is necessary to be able to provide a path to the .config
file without providing the edt.pickle.

The edt.pickle field is now optional in --core parameter.
Binary field is also optional.
If left empty, the utility will not attempt to open the pickle file and
devicetree and binary path information will not be supplied to the
template generator.

Signed-off-by: Rafał Kuźnia <rafal.kuznia@nordicsemi.no>