templates: add retries for streaming operations
Streaming operations, such as calculating the digest or
suit-directive-copy are now executed at least twice.

This increases the robustness of the update process and may prevent the
device from entering the recovery state when a bit flip occurs during
the streaming operation, but the update candidate is correct and
uncorrupted. In such cases, the bit flip is recoverable.
The device will repeat the streaming operation with the hope that the
bit flip does not occur again.
One source of recoverable bit flips may be interference on the SPI bus
when streaming from external memory, or a bit flip when writing
a candidate to NVM.

Signed-off-by: Rafał Kuźnia <rafal.kuznia@nordicsemi.no>