Used pynrfkms instead of command line kms in encrypt_script.py
It is (by the current state of knowledge) impossible to pass
a binary AAD (containing non printable sign) to KMS via the
command line API. The required AAD contains such bytes,
which is why the pynrfkms frontend has to be used.

Note that the limitation of the pynrfkms frontend is that
it cannot used the already initialized KMS backend. This
is why additional parameters have been added to
encrypt_script.py to allow for backend initialization.

Ref: NCSDK-28259

Signed-off-by: Artur Hadasz <artur.hadasz@nordicsemi.no>